Module Dream_html.Path

Bidirectional paths with type-safe path segment parsing and printing using OCaml's built-in format strings, and fully plug-and-play compatible with Dream routes. Refer to Path.make for the syntax of the route strings.

type (_, _) t

This contains the path formats that are used to parse and print the routes.

val make : ('r, unit, Dream.response Dream.promise) Stdlib.format -> ('p, unit, string, attr) Stdlib.format4 -> ('r, 'p) t

make request_fmt attr_fmt is a route path. The dream-html.ppx provides a more convenient way.

Without PPX: let order = Path.make "/orders/%s" "/orders/%s"

With PPX: let order = [%path "/orders/%s"]

Refer to the PPX documentation for instructions on using it.

Due to the way Dream's router works, all parameter captures happens between / characters and the end of the path. Eg, /foo/%s/bar/%d is valid, but /foo/%s.%s (note the dot character) is not a valid capture.

The following type conversion specs are supported:

%s capture a string and pass it to the handler

%*s capture the rest of the path and pass the captured length and string to the handler

%c capture a char

%d or %i capture an int

%x capture a hexadecimal int

%X capture an uppercase hexadecimal int

%o capture an octal int

%ld capture an int32

%Ld capture an int64

%f capture a float

%B capture a bool

val pp : (_, _) t Fmt.t

pp is a pretty-printer for path values. For a path like Path.make "/foo", "/foo", it will print out /foo.